Add base ExoPlayer impl.

+++ player/core/src/main/java/com/coopsrc/oneplayer/core/IOnePlayer.java
@@ -57,7 +57,7 @@ public interface IOnePlayer extends IPlayerInfo {
 
     void seekTo(long msec, int mode);
 
-    void seekTo(int msec) throws IllegalStateException;
+    void seekTo(long msec) throws IllegalStateException;
 
     void setVolume(float volume);
