+++ player/ijkmedia/src/main/java/com/coopsrc/oneplayer/ijkmedia/OneIjkPlayer.java
@@ -187,7 +187,7 @@ public class OneIjkPlayer extends AbsOnePlayer<IjkMediaPlayer> {
     }
 
     @Override
-    public void seekTo(int msec) throws IllegalStateException {
+    public void seekTo(long msec) throws IllegalStateException {
         if (mInternalPlayer != null) {
             mInternalPlayer.seekTo(msec);
         }
